agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
